+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## Bug Fixes
 
 * Assigning a managed object's own list to itself would accidentally clear it (#5395).
+* Don't try to acquire `ApplicationContext` if not available in `Realm.init(Context)` (#5389).
 
 ## Internal